move and rename myDrawCylinderImageFilter in DrawCylinderImageFilter.h
