+++ code/rtkDrawCylinderImageFilter.txx
@@ -23,70 +23,17 @@
 #include <itkImageRegionConstIterator.h>
 #include <itkImageRegionIterator.h>
 
-#include "rtkHomogeneousMatrix.h"
+#include "rtkDrawCylinderImageFilter.h"
 
 namespace rtk
 {
 
-template <class TInputImage, class TOutputImage>
-DrawCylinderImageFilter<TInputImage, TOutputImage>
+template <class TInputImage, class TOutputImage, typename TFunction>
+DrawCylinderImageFilter<TInputImage, TOutputImage, TFunction>
 ::DrawCylinderImageFilter()
-{
-  m_Axis.Fill(90.);
-  m_Center.Fill(0.);
-  m_Density = 1.;
-  m_Angle = 0.;
-}
-
-template <class TInputImage, class TOutputImage>
-void DrawCylinderImageFilter<TInputImage, TOutputImage>::ThreadedGenerateData(const OutputImageRegionType& outputRegionForThread,
-                                                                              ThreadIdType itkNotUsed(threadId) )
-{
-  //Getting phantom parameters
-  EQPFunctionType::Pointer sqpFunctor = EQPFunctionType::New();
-  FigureType Cylinder;
-
-  Cylinder.semiprincipalaxis[0] = m_Axis[0];
-  Cylinder.semiprincipalaxis[1] = m_Axis[1];
-  Cylinder.semiprincipalaxis[2] = m_Axis[2];
-  Cylinder.center[0] = m_Center[0];
-  Cylinder.center[1] = m_Center[1];
-  Cylinder.center[2] = m_Center[2];
-  Cylinder.angle = m_Angle;
-  Cylinder.density = m_Density;
-
-  typename TOutputImage::PointType point;
-  const    TInputImage *           input = this->GetInput();
-
-  typename itk::ImageRegionConstIterator<TInputImage> itIn( input, outputRegionForThread);
-  typename itk::ImageRegionIterator<TOutputImage> itOut(this->GetOutput(), outputRegionForThread);
-
-  //Set type of Figure
-  sqpFunctor->SetFigure("Cylinder");
-  //Translate from regular expression to quadric
-  sqpFunctor->Translate(Cylinder.semiprincipalaxis);
-  //Apply rotation and translation if necessary
-  sqpFunctor->Rotate(Cylinder.angle, Cylinder.center);
-
-  while( !itOut.IsAtEnd() )
-  {
-    this->GetInput()->TransformIndexToPhysicalPoint(itOut.GetIndex(), point);
-
-    double QuadricEllip = sqpFunctor->GetA()*point[0]*point[0]   +
-                          sqpFunctor->GetB()*point[1]*point[1]   +
-                          sqpFunctor->GetC()*point[2]*point[2]   +
-                          sqpFunctor->GetD()*point[0]*point[1]   +
-                          sqpFunctor->GetE()*point[0]*point[2]   +
-                          sqpFunctor->GetF()*point[1]*point[2]   +
-                          sqpFunctor->GetG()*point[0] + sqpFunctor->GetH()*point[1] +
-                          sqpFunctor->GetI()*point[2] + sqpFunctor->GetJ();
-    if(QuadricEllip<0)
-      itOut.Set(Cylinder.density );
-    else
-      itOut.Set( itIn.Get() );
-    ++itIn;
-    ++itOut;
-  }
+{  
+  this->m_spatialObject.sqpFunctor = EQPFunctionType::New();
+  this->m_spatialObject.sqpFunctor->SetFigure("Cylinder");
 }
 
 }// end namespace rtk
