+++ code/rtkDrawImageFilter.h
@@ -114,101 +114,6 @@ class ITK_EXPORT DrawImageFilter :
   };
   
   
-  template <class TInputImage, 
-            class TOutputImage, 
-	    typename TFunction = itk::Functor::Add2<typename TInputImage::PixelType,
-                                                    typename TInputImage::PixelType,
-                                                    typename TOutputImage::PixelType>
-                                                    >
-  class myDrawCylinderImageFilter : 
-  public DrawImageFilter< TInputImage,
-                         TOutputImage, 
-			 DrawCylinderSpatialObject,
-			 TFunction
-			 >
-  {
-       public:
-  /** Standard class typedefs. */
-  typedef myDrawCylinderImageFilter                               Self;
-  typedef itk::InPlaceImageFilter<TInputImage,TOutputImage> Superclass;
-  typedef itk::SmartPointer<Self>                           Pointer;
-  typedef itk::SmartPointer<const Self>                     ConstPointer;
-  typedef typename TOutputImage::RegionType                 OutputImageRegionType;
-  
-  typedef itk::Vector<double,3>                             VectorType;
-  typedef rtk::ConvertEllipsoidToQuadricParametersFunction  EQPFunctionType;
-  
-  struct FigureType
-  {
-    FigureType():angle(0.),density(0.){};
-    VectorType semiprincipalaxis;
-    VectorType center;
-    double     angle;
-    double     density;
-  }; 
-  
-  
-  void SetAxis(VectorType Axis)
-  {
-    if ( Axis == this->m_spatialObject.m_Axis )
-      {
-      return;
-      }
-    this->m_spatialObject.m_Axis = Axis;
-    this->m_spatialObject.sqpFunctor->Translate(this->m_spatialObject.m_Axis);
-    this->m_spatialObject.sqpFunctor->Rotate(this->m_spatialObject.m_Angle, this->m_spatialObject.m_Center);
-    this->Modified();
-  }  
-  
-  
-   void SetCenter(VectorType Center)
-  {
-    if ( Center == this->m_spatialObject.m_Center )
-      {
-      return;
-      }
-    this->m_spatialObject.m_Center = Center;
-    this->m_spatialObject.sqpFunctor->Translate(this->m_spatialObject.m_Axis);
-    this->m_spatialObject.sqpFunctor->Rotate(this->m_spatialObject.m_Angle, this->m_spatialObject.m_Center);
-    this->Modified();
-  }
-  
-  void SetAngle(double Angle)
-  {
-    if ( Angle == this->m_spatialObject.m_Angle )
-      {
-      return;
-      }
-    this->m_spatialObject.m_Angle = Angle;
-    this->m_spatialObject.sqpFunctor->Translate(this->m_spatialObject.m_Axis);
-    this->m_spatialObject.sqpFunctor->Rotate(this->m_spatialObject.m_Angle, this->m_spatialObject.m_Center);
-    this->Modified();
-  }
-  
-  
-  
-  
-  /** Method for creation through the object factory. */
-  itkNewMacro(Self);
-
-  /** Run-time type information (and related methods). */
-  itkTypeMacro(myDrawCylinderImageFilter, DrawImageFilter);
-  
-  protected:
-  myDrawCylinderImageFilter();
-  virtual ~myDrawCylinderImageFilter() {};
-//   virtual void ThreadedGenerateData( const OutputImageRegionType& outputRegionForThread, ThreadIdType threadId ) ITK_OVERRIDE;
-    
-  
-  
-  
-  private:
-      myDrawCylinderImageFilter(const Self&); //purposely not implemented
-      void operator=(const Self&);            //purposely not implemented
- 
-  
-  
-  }; 
 
   
 } // end namespace rtk
